extract helper class from runtime handler
